Remove wasm feature in favor of being compatible by default
