+++ src/cli/new_transactions.rs
@@ -19,6 +19,7 @@ use miden_objects::{
 };
 use miden_tx::auth::TransactionAuthenticator;
 use tracing::info;
+use winter_maybe_async::maybe_await;
 
 use super::{
     get_input_note_with_id_prefix,
@@ -246,7 +247,7 @@ impl ConsumeNotesCmd {
             .list_of_notes
             .iter()
             .map(|note_id| {
-                get_input_note_with_id_prefix(client, note_id)
+                maybe_await!(get_input_note_with_id_prefix(client, note_id))
                     .map(|note_record| note_record.id())
                     .map_err(|err| err.to_string())
             })
