+++ src/client/sync.rs
@@ -574,9 +574,8 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
         Ok(PartialMmr::from_parts(current_peaks, tracked_nodes, track_latest))
     }
 
-    #[cfg_attr(feature = "wasm", allow(rustdoc::broken_intra_doc_links))]
     /// Extracts information about nullifiers for unspent input notes that the client is tracking
-    /// from the received [SyncStateResponse]
+    /// from the received list of nullifiers in the sync response
     #[maybe_async]
     fn get_new_nullifiers(&self, new_nullifiers: Vec<Digest>) -> Result<Vec<Digest>, ClientError> {
         // Get current unspent nullifiers
