+++ src/client/transactions/mod.rs
@@ -5,7 +5,6 @@ use miden_objects::{
     accounts::{AccountDelta, AccountId, AuthSecretKey},
     assembly::ProgramAst,
     assets::FungibleAsset,
-    crypto::rand::RpoRandomCoin,
     notes::{Note, NoteDetails, NoteId, NoteType},
     transaction::{
         ExecutedTransaction, InputNote, InputNotes, OutputNote, OutputNotes, ProvenTransaction,
@@ -14,10 +13,6 @@ use miden_objects::{
     Digest, Felt, FieldElement, Word,
 };
 use miden_tx::{auth::TransactionAuthenticator, ProvingOptions, ScriptTarget, TransactionProver};
-#[cfg(not(feature = "wasm"))]
-use rand::Rng;
-#[cfg(feature = "wasm")]
-use rand::{rngs::StdRng, Rng, SeedableRng};
 use tracing::info;
 use winter_maybe_async::{maybe_async, maybe_await};
 
@@ -327,37 +322,16 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
     // HELPERS
     // --------------------------------------------------------------------------------------------
 
-    /// Gets [RpoRandomCoin] from the client
-    #[cfg(not(feature = "wasm"))]
-    fn get_random_coin(&self) -> RpoRandomCoin {
-        // TODO: Initialize coin status once along with the client and persist status for retrieval
-        let mut rng = rand::thread_rng();
-        let coin_seed: [u64; 4] = rng.gen();
-
-        RpoRandomCoin::new(coin_seed.map(Felt::new))
-    }
-
-    #[cfg(feature = "wasm")]
-    fn get_random_coin(&self) -> RpoRandomCoin {
-        // TODO: Initialize coin status once along with the client and persist status for retrieval
-        let mut rng = StdRng::from_entropy();
-        let coin_seed: [u64; 4] = rng.gen();
-
-        RpoRandomCoin::new(coin_seed.map(Felt::new))
-    }
-
     /// Helper to build a [TransactionRequest] for P2ID-type transactions easily.
     ///
     /// - auth_info has to be from the executor account
     /// - If recall_height is Some(), a P2IDR note will be created. Otherwise, a P2ID is created.
     fn build_p2id_tx_request(
-        &self,
+        &mut self,
         payment_data: PaymentTransactionData,
         recall_height: Option<u32>,
         note_type: NoteType,
     ) -> Result<TransactionRequest, ClientError> {
-        let mut random_coin = self.get_random_coin();
-
         let created_note = if let Some(recall_height) = recall_height {
             create_p2idr_note(
                 payment_data.account_id(),
@@ -366,7 +340,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
                 note_type,
                 Felt::ZERO,
                 recall_height,
-                &mut random_coin,
+                &mut self.rng,
             )?
         } else {
             create_p2id_note(
@@ -375,7 +349,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
                 vec![payment_data.asset()],
                 note_type,
                 Felt::ZERO,
-                &mut random_coin,
+                &mut self.rng,
             )?
         };
 
@@ -414,12 +388,10 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
     ///
     /// - auth_info has to be from the executor account
     fn build_swap_tx_request(
-        &self,
+        &mut self,
         swap_data: SwapTransactionData,
         note_type: NoteType,
     ) -> Result<TransactionRequest, ClientError> {
-        let mut random_coin = self.get_random_coin();
-
         // The created note is the one that we need as the output of the tx, the other one is the
         // one that we expect to receive and consume eventually
         let (created_note, payback_note_details) = create_swap_note(
@@ -428,7 +400,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
             swap_data.requested_asset(),
             note_type,
             Felt::ZERO,
-            &mut random_coin,
+            &mut self.rng,
         )?;
 
         let recipient = created_note
@@ -466,19 +438,18 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
     ///
     /// - faucet_auth_info has to be from the faucet account
     fn build_mint_tx_request(
-        &self,
+        &mut self,
         asset: FungibleAsset,
         target_account_id: AccountId,
         note_type: NoteType,
     ) -> Result<TransactionRequest, ClientError> {
-        let mut random_coin = self.get_random_coin();
         let created_note = create_p2id_note(
             asset.faucet_id(),
             target_account_id,
             vec![asset.into()],
             note_type,
             Felt::ZERO,
-            &mut random_coin,
+            &mut self.rng,
         )?;
 
         let recipient = created_note
